requirements.txt
django>=3.2.1,<3.3
dj-database-url
python-decouple
django-uuid-upload
Pillow
django-solo
django-python3-ldap
django-minio-storage
psycopg2-binary
gunicorn
django-currentuser
django-ipware
django-session-timeout
whitenoise
zoomus
python-dateutil
